+++ packages/core/src/shipping/consignment-action-creator.ts
@@ -2,6 +2,8 @@ import { createAction, createErrorAction, ThunkAction } from '@bigcommerce/data-
 import { find } from 'lodash';
 import { Observable, Observer } from 'rxjs';
 
+import { isExperimentEnabled } from '@bigcommerce/checkout-sdk/utility';
+
 import { AddressRequestBody } from '../address';
 import { Cart } from '../cart';
 import {
@@ -358,6 +360,26 @@ export default class ConsignmentActionCreator {
         return (store) =>
             Observable.create((observer: Observer<UpdateShippingOptionAction>) => {
                 const checkout = store.getState().checkout.getCheckout();
+                const { checkoutSettings } = store.getState().config.getStoreConfigOrThrow();
+
+                if (
+                    isExperimentEnabled(
+                        checkoutSettings.features,
+                        'CHECKOUT-8999.remove_duplicate_shipping_option_call',
+                    )
+                ) {
+                    const consignmentInMemory = store
+                        .getState()
+                        .consignments.getConsignmentById(consignment.id);
+
+                    const alreadySelectedOptionId = consignmentInMemory?.selectedShippingOption?.id;
+
+                    if (alreadySelectedOptionId === consignment.shippingOptionId) {
+                        observer.complete();
+
+                        return;
+                    }
+                }
 
                 if (!checkout || !checkout.id) {
                     throw new MissingDataError(MissingDataErrorType.MissingCheckout);