+++ packages/core/src/shipping/consignment.ts
@@ -49,6 +49,7 @@ export interface ConsignmentUpdateRequestBody {
     shippingAddress?: AddressRequestBody;
     lineItems?: ConsignmentLineItem[];
     pickupOption?: ConsignmentPickupOption;
+    shippingOptionId?: string;
 }
 
 export interface ConsignmentMeta {