+++ packages/google-pay-integration/src/google-pay-paypal-commerce/google-pay-paypal-commerce-script-loader.ts
@@ -79,6 +79,8 @@ export default class GooglePayPayPalCommerceScriptLoader {
         return this.window.paypal;
     }
 
+    // TODO: Remove this rule disabling after method refactor
+    // eslint-disable-next-line complexity
     private getPayPalSdkScriptConfigOrThrow(
         paymentMethod: PaymentMethod<GooglePayPayPalCommerceInitializationData>,
         currencyCode: string,