+++ packages/utility/src/is-experiment-enabled/index.ts
@@ -1,0 +1,1 @@
+export { default as isExperimentEnabled } from './is-experiment-enabled';